Replace satori/go.uuid by gofrs/uuid
The satori/go.uuid package has critical
bugs (satori/go.uuid#73).

This commit replaces this package by the gofrs/uuid fork.

In that package, The Equal function for UUID has been deleted (we
can use == directly for comparisons as explained in the 3.0.0
release note of the package:
https://github.com/gofrs/uuid/releases/tag/v3.0.0).

Related to containous/traefik#4718