Makefile: fix targets (#619)
Due to the introduction of `.ONESHELL` to `Makefile` in #618 the shell
commands in each target are now executed in one shell session. This
improves performance but also means that changing the directory in one
line also changes the directory for the following lines(doesn't happen
across targets). Unfortunately it broke the targets `oapigen` and
`generate`.